major bug that crushed my project, changed pupil dao

+++ pupil-service/src/main/java/com/example/pupilservice/dto/PupilDTO.java
@@ -25,7 +25,7 @@ public class PupilDTO {
     private String email;
     private String username;
     private String avatarId;
-    private Integer level;
+    private double level;
     private double coins;
